pyproject.toml
[build-system]
requires = ["setuptools"]
build-backend = "setuptools.build_meta"

[project]
name = "mds_mapper"
dynamic = ["version"]
authors = [{ name = "Joshua Lauterbach", email = "jl@beebucket.de" }]
description = "A small service to convert Beebucket EDP's into the MDS Ontology"
readme = "README.md"
requires-python = ">=3.12"
classifiers = [
    "Development Status :: 3 - Alpha",
    "Natural Language :: English",
    "Operating System :: OS Independent",
    "Programming Language :: Python",
    "Programming Language :: Python :: 3.12",
]
dependencies = [
    "extended-dataset-profile @ git+https://github.com/Beebucket/extended_dataset_profile@0.26.1",
    "pg8000",
    "pydantic",
    "pydantic-settings",
    "python-iso639",
    "requests",
    "sqlmodel",
    "fastapi[standard]",
]

[project.scripts]
mds_mapper_rest_api = "mds_mapper.rest_api.__main__:main"

[project.optional-dependencies]
test = [
    "pytest",
    "pytest-asyncio",
    "mypy",
    "types-requests",
    "ruff>=0.7.4",
    "pre-commit>=4.0.1",
    "pip-licenses>=5.0.0",
    "pydantic-markdown @ git+https://github.com/Beebucket/pydantic_markdown@2.0.1",
]
dev = ["mds_mapper[test]"]

[tool.setuptools.packages.find]
where = ["src"]

[tool.setuptools.dynamic]
version = { file = "dist/VERSION.txt" }

[tool.ruff]
line-length = 120

[tool.ruff.lint]
extend-select = [
    "I", # isort replacement
    "S", # bandit
]

[tool.ruff.lint.per-file-ignores]
"tests/*" = ["S101", "S106"]


[tool.pytest.ini_options]
log_cli = true
log_cli_level = "WARNING"
log_format = "%(asctime)s | %(levelname)s | %(name)s | %(pathname)s:%(lineno)d | - %(message)s"
asyncio_mode = "auto"
asyncio_default_fixture_loop_scope = "session"
filterwarnings = ["error"]
markers = ["slow: Tests that may take more than a few seconds."]

[tool.mypy]
warn_return_any = true
warn_unused_configs = true
exclude = ["build", ".venv"]
check_untyped_defs = true
sqlite_cache = true
cache_fine_grained = true

[tool.pip-licenses]
from = "classifier"
ignore-packages = ["extended-dataset-profile"]
partial-match = true
